Fix cloud shell notebook link
Link to jupyter notebook was not showing up correctly in cloud shell due
to an incorrectly indexed column, which led to the status being stored
as the external_ip rather than the actual external ip. This commit
simply changes the number of the indexed column from '6' to '5'. This
should result in the correct notebook link being printed to the
terminal.